+++ gamit_app/inventory/templatetags/inventory_tags.py
@@ -1,0 +1,14 @@
+from django import template
+
+register = template.Library()
+
+@register.simple_tag(takes_context=True)
+def query_transform(context, **kwargs):
+    request = context['request']
+    updated = request.GET.copy()
+    for k, v in kwargs.items():
+        if v is not None:
+            updated[k] = v
+        else:
+            updated.pop(k, None)
+    return updated.urlencode()